fix: add check for no states matching search term

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -34,9 +34,10 @@ if __name__ == '__main__':
     session = Session()
 
     states_with_letter_a = session.query(State).filter(
-        State.name.like('%a%')).order_by(State.id).all()
+        State.name.contains('a'))
 
-    for instance in states_with_letter_a:
-        session.delete(instance)
+    if states_with_letter_a is not None:
+        for instance in states_with_letter_a:
+            session.delete(instance)
 
     session.commit()
